[#122] feat: 하단 텍스트필드 디자인 적용
