+++ lib/features/swipe_view/presentation/screen/swipe_view.dart
@@ -131,60 +131,6 @@ class SwipeViewState extends ConsumerState<SwipeView>
                           ),
                         ),
                       ),
-                      Column(
-                        children: [
-                          // SizedBox(
-                          //   height: _swipeViewModel.commentFieldFocus.hasFocus
-                          //       ? 0
-                          //       : 50,
-                          //   child: Row(
-                          //     children: [
-                          //       Expanded(child: Container()),
-                          //       GestureDetector(
-                          //         onTapUp: (details) async =>
-                          //             emojiSheetWidget(context)
-                          //                 .whenComplete(() async {
-                          //           _swipeViewModelProvider.sendEmojiReaction();
-                          //         }),
-                          //         child: Container(
-                          //           width: 50,
-                          //           height: 50,
-                          //           decoration: const BoxDecoration(
-                          //             shape: BoxShape.circle,
-                          //           ),
-                          //           child: const Center(
-                          //             child: Text(
-                          //               '😄',
-                          //               style: TextStyle(fontSize: 30),
-                          //             ),
-                          //           ),
-                          //         ),
-                          //       ),
-                          //       Container(
-                          //         key: _swipeViewModel.cameraButtonPlaceholderKey,
-                          //         width: 50,
-                          //         height: 50,
-                          //         decoration: const BoxDecoration(
-                          //           shape: BoxShape.circle,
-                          //           // color: Colors.black,
-                          //         ),
-                          //         child: Center(
-                          //           child: Text(
-                          //             _swipeViewModel.isCameraInitialized == true
-                          //                 ? '📸'
-                          //                 : '❌',
-                          //             style: const TextStyle(fontSize: 30),
-                          //           ),
-                          //         ),
-                          //       ),
-                          //     ],
-                          //   ),
-                          // ),
-                          // const SizedBox(
-                          //   height: 80,
-                          // ),
-                        ],
-                      ),
                     ],
                   ),
                 ),
@@ -201,102 +147,6 @@ class SwipeViewState extends ConsumerState<SwipeView>
     );
   }
 
-  Future<dynamic> emojiSheetWidget(BuildContext context) {
-    return showModalBottomSheet(
-      backgroundColor: Colors.transparent,
-      clipBehavior: Clip.none,
-      elevation: 0,
-      context: context,
-      builder: (context) {
-        void disposeWidget(UniqueKey key) {
-          setState(() {
-            _swipeViewModel.emojiWidgets.remove(key);
-          });
-        }
-
-        return StatefulBuilder(
-          builder: (BuildContext context, StateSetter setState) {
-            return Stack(
-              alignment: Alignment.center,
-              clipBehavior: Clip.none,
-              children: [
-                Stack(
-                  alignment: Alignment.topCenter,
-                  clipBehavior: Clip.none,
-                  children: _swipeViewModel.emojiWidgets.values.toList(),
-                ),
-                Container(
-                  clipBehavior: Clip.hardEdge,
-                  decoration: const BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.only(
-                      topLeft: Radius.circular(20),
-                      topRight: Radius.circular(20),
-                    ),
-                  ),
-                ),
-                Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: <Widget>[
-                    Padding(
-                      padding: const EdgeInsets.symmetric(vertical: 30.0),
-                      child: Text(
-                        '반응을 ${_swipeViewModel.countSend}회 보냈어요!',
-                        style: const TextStyle(fontSize: 20),
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                      child: Column(
-                        children: List<Widget>.generate(
-                          3,
-                          (index) => Row(
-                            children: List<Widget>.generate(5, (jndex) {
-                              final key = GlobalKey();
-                              return Expanded(
-                                key: key,
-                                child: GestureDetector(
-                                  onTapDown: (detail) {},
-                                  onTapUp: (detail) {
-                                    shootEmoji(
-                                      setState,
-                                      index,
-                                      jndex,
-                                      detail,
-                                      context,
-                                      disposeWidget,
-                                    );
-                                  },
-                                  child: Stack(
-                                    clipBehavior: Clip.none,
-                                    children: [
-                                      Image(
-                                        image: AssetImage(
-                                          Emojis.emojiList[index * 5 + jndex],
-                                        ),
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                              );
-                            }),
-                          ),
-                        ),
-                      ),
-                    ),
-                    const SizedBox(
-                      height: 60,
-                    ),
-                  ],
-                ),
-              ],
-            );
-          },
-        );
-      },
-    );
-  }
-
   void setAnimationVariables() {
     _swipeViewModel.animationController = AnimationController(
       vsync: this,
@@ -309,37 +159,5 @@ class SwipeViewState extends ConsumerState<SwipeView>
       ),
     );
     _swipeViewModel.animationController.value = 1;
-  }
-
-  void shootEmoji(
-    StateSetter setState,
-    int index,
-    int jndex,
-    TapUpDetails detail,
-    BuildContext context,
-    void Function(UniqueKey key) disposeWidget,
-  ) {
-    return setState(
-      () {
-        _swipeViewModel.countSend++;
-        _swipeViewModel.sendingEmojis[index * 5 + jndex] += 1;
-        final animationWidgetKey = UniqueKey();
-        _swipeViewModel.emojiWidgets.addEntries(
-          {
-            animationWidgetKey: ShootEmojiWidget(
-              key: animationWidgetKey,
-              emojiIndex: index * 5 + jndex,
-              currentPos:
-                  Point(detail.globalPosition.dx, detail.globalPosition.dy),
-              targetPos: Point(
-                MediaQuery.of(context).size.width / 2,
-                MediaQuery.of(context).size.height + 50,
-              ),
-              disposeWidgetFromParent: disposeWidget,
-            ),
-          }.entries,
-        );
-      },
-    );
   }
 }
